added(logout): Set view to chat after logout (CODY-3606) (#5544)
[CODY-3606](https://linear.app/sourcegraph/issue/CODY-3606/first-time-login-redirects-to-profile-page-instead-of-chat-page)

If the user signs out and then signs back in, the view is still cached
as the settings page. This updates the view to the chat page before
signout.

I tried to implement this in the ChatController but could not get it to
work, and sank an embarrassing amount of time before finding [this
sneaky little special
case](https://sourcegraph.com/github.com/sourcegraph/jetbrains/-/blob/src/main/kotlin/com/sourcegraph/cody/ui/web/WebUIHost.kt?L71%3A1-74%3A1).

Also added a proxy to log messages from the webview.

## Test plan
Manually tested. I could also add an integration test if we think it's a
good idea.