expose ModelsService.changes and listen to that in ChatController (#5578)
Previously, the ChatController listened for auth changes and assumed
that model changes would have been already applied. This required a
specific ordering of operations in the AuthProvider, which was a brittle
way of ensuring this behavior (and the `chat model selector` e2e test
broke when that behavior was slightly changed). Now, the ModelsService
exposes a `changes` observable and the ChatController listens to that to
follow changes to models, which makes sense.

## Test plan

e2e tests cover this well.