Testing the Removal of Artificial Latency behind a feature flag (#5480)
This PR aims to remove the artificial delay introduced for
low-performance languages and specific AST nodes, as outlined in th
[issue](https://linear.app/sourcegraph/issue/CODY-3315/delete-the-artificial-latency).
The artificial delay was originally implemented for low performance
languages so that users are not spammed with unhelpful suggestions. ,
but it has become less relevant over time, and it may now impact the
predictability of autocomplete latency and user experience (UX).

# Changes:

## Introduce a feature flag:
A new feature flag has been added to toggle the artificial delay.
- Objective: The main objective is to achieve predictable latency and a
consistent autocomplete UX for all languages.

- After merging this PR, the feature flag will be enabled for controlled
testing, and data will be collected to assess the impact on CAR and
latency for all languages
- Depending on the results, the team will decide whether to remove the
artificial delay logic entirely.

## Test Plan:
I tested this PR locally by enabling and disabling the feature flag on
my local Sourcegraph instance(I added the feature flag on it and then
set the true/false values). I also used the debugger to run VS Code
Extension on Desktop



## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->