Autocomplete: expose and test request params (#5604)
The problem: currently, we
[assert](https://github.com/sourcegraph/cody/blob/d10b88f3bcaad1c140352088a15f641f4697c80f/vscode/src/completions/providers/shared/create-provider.test.ts#L218-L219)
the model ID assigned to the provider instance, but often, it's not used
at all to make requests, which means this assertion is irrelevant to the
actual functionality where we bump into bugs.

- This PR adds tests for what matters — autocomplete request parameters
in all supported instance configurations: dotcom/enterprise X
local-settings/server-side-model-config/site-config.
- Adds several test helpers to ease the addition of similar tests to all
the autocomplete providers. Currently we assert request params only for
the `unstable-openai` provider. Will do the rest in a follow-up PR.
- Follow up for https://github.com/sourcegraph/cody/pull/5595.
- No functional changes.