Fix cody bench and add eval strategy for chat context (#5624)
This fixes `cody internal bench`, which was broken by some recent
changes on how config is synced. It ensures `localStorage` using an
in-memory store before its use in the bench command.

It also adds a new eval strategy for context that consumes a CSV of
queries and computes eval metrics for context quality.