disable fetching Ollama models during vitest unit testing to avoid nondeterminism (#5657)
This will make our tests more reliable when run locally.

It is possible that this could make it harder to spot issues in unit
tests and agent tests where Ollama being available causes issues. If we
see such an issue, then we should build in a case to prevent that
problem. As is, CI and many/most devs aren't running Ollama, so we have
at best uneven protection, and at worst the flakiness causes wasted dev
time and possibly bad workarounds.

## Test plan

CI