Use `ps` for PromptString literals. (#5607)
Do not use PromptString.unsafe_fromLLMResponse for anything other than
LLM responses. In this case, the string is a literal and should just be
constructed with `ps`.

Remove the mention of VSCode. This string shows up in all IDEs, so
"VSCode settings" is confusing in JetBrains, Visual Studio, Eclipse.

## Test plan

CI

---------

Co-authored-by: Naman Kumar <naman@outlook.in>