Remove v1 telemetry from JB extension (#2307)
So I don't think that the CI error (LSIF index generation) is related to
this PR, so I'm opening this up for review.

This PR is intended to remove legacy ("V1") telemetry from the JetBrains
extension. As I'm highly unfamiliar with the codebase, I'd love a review
from anyone who is a bit more familiar with this! Feel free to reassign
to someone else or ask questions @pkukielka

## Test plan

CI

(Unfortunately I don't have a JB license. I can get one if you think
this change needs a live test)

---------

Co-authored-by: Dan Adler <5589410+dadlerj@users.noreply.github.com>