fix switch account bug
The problem is that the global resolvedConfig observable (because of `shareReplay`) continued to emit the previous value when the underlying config changed, because it has a brief async step to fetch the secrets from the secret store. During that period (which does not require network access but is more than a single tick), it would emit the old auth endpoint and access token, which meant that stuff happening in the brief period after switching accounts would sometimes use the wrong endpoint and auth token, thereby getting incorrect values. (It never sent the wrong access token to the wrong endpoint.)

The fix is to use switchMapReplayOperation, which is a wrapper around switchMap that is used elsewhere. This basically means: "memoize the output of this operation, but immediately stop returning the old data when there's a new input until we complete the operation on that input". That's what we use a lot of other places, and I forgot to update it here to add it.