observe config changes in AuthProvider (#5221)
The goal is to make Cody respond to config and auth changes better, so
that no full-editor-reloads are required, and it reacts instantly to
connectivity loss and reestablishment. This PR is one of many related
ones, and this particular one makes it so that the AuthProvider listens
to config changes more reactively. It's still not totally reactive;
sometimes `AuthProvider.auth` still is called. But it gets us closer.

Helps with
https://linear.app/sourcegraph/issue/CODY-3332/full-native-offline-mode-and-instant-config-reactivity-for-cody-in-vs.

Original intent:
https://sourcegraph.slack.com/archives/C05AGQYD528/p1723791822404399.

## Known issues

We will need to fix these before releasing this:

There are still some glitches around account switching and getting the
client config (ClientConfigSingleton). Almost all of these issues are
related to the clientConfig observable being out of sync with the
authStatus for a tick, which I haven’t figured out the reason for but
have plenty of ideas.

- [ ] When switching from dotcom to s2, the model selector does not show
up post-switch in any visible chats
- [ ] When switching from s2 to dotcom, it shows
non-server-side-model-config models for 60 seconds until it refreshes
client config
- [ ] When switching from dotcom to s2, it briefly shows the "could not
create 'sourcegraph' autocomplete provider" message but then
autocomplete does work because it immediately fixes its configuration
- [ ] On dotcom on a freshly reloaded editor, sometimes when you chat,
it says it will use Claude 3.5 Sonnet but actually uses an Ollama model.
- [ ] I want to double-check that the embeddings service is only running
for dotcom.
- [ ] And that the embeddings service starts up when you switch from s2
to dotcom, and shuts down when you switch from dotcom to s2.
- [ ] The ClientConfigSingleton now fetches new client config every 60s.
Previously it would trigger an async refetch if the data was more than
60s old and was requested due to a user action. This means that it will
now fetch in the background more often, which reintroduces the problem
https://linear.app/sourcegraph/issue/CODY-3745/codys-background-periodic-network-access-causes-2fa.
We need to make it not perform its background fetch if the editor is not
the foreground window.
- [ ] Switching models is slow. There is probably some kind of state
update cycle. Need to debug this by adding `tapLog` pipe operators in
some observable chains.
- [ ] `Error: command 'cody.embeddings.resolveIssue' already exists`
- [ ] When the editor is first starting, a chat editor panel briefly
shows the login screen (probably before the auth is finished) and then
shows chat after auth is finished. Need to make the chat wait for some
kind of initialization before showing anything.
- [ ] Need to confirm that SiteProductVersion request spam is not
occurring. (Have confirmed but want to do so again before a release in
case anything changes.)


## Test plan

e2e tests, and the account switching described above.

---------

Co-authored-by: Vova Kulikov <vova@sourcegraph.com>