make unified prompts feature flag reactive (#5672)
Fixes an issue where the auth was not yet available during init, so
`evaluateFeatureFlag` could not be called in Cody Web (which supplies
creds after initialization). Also makes it so that an editor reload is
not necessary to pick up an updated value for this feature flag. (I will
remove the `evaluateFeatureFlag` method because it is hard to think of a
situation where you only want to read a feature flag's value at a given
instant and not watch its value.)

## Test plan

n/a