Agent: clean-up assertion of telemetry events within integration tests (#5455)
Previously, the assertion of telemetry events for agent integration
tests was happening within an `afterEach`, and this was not the cleanest
approach, as the error messages would not point to the relevant code
causing the error. In addition, this PR also adds `parameters` as a
property that gets sent with the testing telemetry events.

Now, we manually add the assertion at the end of test cases where we
want to assert what events got exported. As @olafurpg mentioned, the
benefits of this approach are:

> The benefit of this approach is that you get clearer error messages
pointing to a relevant line of code in the tests. With the afterEach()
approach, you need to hunt down the test name and reason about this
currentClient logic. Also, this afterEach() block is only available in
index.test.ts, it's not straightforward how to reuse this logic in other
test files.

Addresses feedback
[here](https://github.com/sourcegraph/cody/pull/4923#discussion_r1722970638)

## Test plan
CI, specifically `pnpm run test src/index.test.ts`
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->