remove "enhanced context" concept
The concept of "enhanced context" was removed in VS Code in May 2024 and in JetBrains in early Sep 2024 when we unified the chat UI across all editors. Now, the chat message explicitly @-mentions repositories that should be searched for context. There were still many remnants of "enhanced context" in our internal and agent APIs that were no longer being used. Removing them simplifies the code.