treat ChatBuilder (nee ChatModel) as having minimal state
Previously, the ChatModel was a source of truth and a cache. It cached the model that would be used for the next assistant message, and ChatController had to update that whenever the default chat model changed. Now, ChatBuilder (renamed from ChatModel to be less confusing) only stores the chat model *if* the user explicitly selected one; otherwise, its `model` is `undefined` and it is resolved to the default chat model as late as possible, when the chat is actually submitted. The actual chat model used is then stored in the message transcript.

This is better because ChatController does not need to worry about stale state being present in the ChatBuilder, except if the user's explicitly chosen model is no longer available on the endpoint, which is an error case that needs to be handled interactively (and isn't an error case that is just present because of our internal logic).

Closes https://linear.app/sourcegraph/issue/CODY-3845/fix-chatcontroller-chipmunk-state.