do not update ClientConfig when the editor is in the background
The ClientConfigSingleton was changed last week to fetch new client config every 60s. Previously it would trigger an async refetch if the data was more than 60s old and was requested due to a user action. This means that it would now fetch in the background, which reintroduces the problem https://linear.app/sourcegraph/issue/CODY-3745/codys-background-periodic-network-access-causes-2fa. This change fixes that again in this new implementation.

Fixes https://linear.app/sourcegraph/issue/CODY-3844/make-clientconfigsingleton-not-update-when-the-editor-is-in-the.