fix explicitly selected model on new chat is not reset on account switch (#5715)
If you are on endpoint E1 and choose a model M1 in a new chat that is
NOT sent, and then switch accounts to endpoint E2, the model selector
will show your default model for E2 but your unsent chat will, when
sent, use the model M1, which is likely not valid on E2. Now, we reset
the new chat when your endpoint changes, which eliminates this problem.

Fixes
https://linear.app/sourcegraph/issue/CODY-3898/explicitly-selected-model-in-new-chat-is-not-overwritten-on-account.

## Test plan

As described above.