rename gitUrl concept to repoName, avoid there-and-back transformation
In autocomplete we log the repo name (a string like `github.com/foo/bar`). We were calling this `gitUrl`, but typically in Sourcegraph-land we call this the "repo name". In this code, we were needlessly passing around an actual or synthesized Git remote URL and then converting it to the repo name, but we actually could just get the repo name directly. Doing so cleans up the code, removes a concept, and makes it so that our repo name mapper needs to expose 1 less public API method.