make clientStateBroadcaster use observables (#5702)
Previously, clientStateBroadcaster used ad-hoc debouncing and observing.
Now, it uses observables, which makes it cleaner. It is also exposed
through the webview API in a simple way instead of having its own
`clientState` context and messages.

## Test plan

Select a file and a range within a file and ensure that the chat message
field updates with that context. Add a `tapLog('initialContext')` to the
`observeInitialContext` observable chain and ensure it does not emit too
often.