simplify repo name->ID mapping (#5717)
There was a lot of complexity in the following classes:

* RepoNameResolver: maps FS paths and git remote URLs to repo names
* WorkspaceRepoMapper: maps repo names to Sourcegraph instance GraphQL
repo IDs
* WorkspaceReposMonitor: combines the above 2 and exposes a list of all
(GraphQL repo ID, commit) values for all workspace folders

It turns out we only really needed 2 things:

* `getRepoNamesContainingUri(uri: vscode.Uri)`
* `remoteReposForAllWorkspaceFolders: Observable<RemoteRepo[]>`

There were a lot of convoluted transformations happening in the code,
such as converting a document URI to a workspace root folder to a Git
remote URL and back to one of those, and then to a repo name and then
repo ID. This simplifies the code.

It also makes it so that all caches are properly evicted when the user
switches accounts. Previously, the caches were not always being cleared,
and old GraphQL repo IDs for different Sourcegraph endpoints would be
stored. Now, caches are cleared correctly, and there is only one cache
and it's basically a memoized `graphqlClient.getRepoName` (previously
there were 4 different and redundant caches).

## Test plan

Test switching between s2 and dotcom while trying:

- Initial context: the repository initial context @-mention should show
up.
- OpenCtx providers: selecting one triggers a call to get the current
repo, and that should not fail
- Remote repository @-mentions on s2

Test autocomplete telemetry: check the output channel logs. Ensure that
the GitHub API or GraphQL API are not being called per keystroke or file
switch (once the values are cached).

Test chat telemetry. Ensure that for public repos, the context is being
logged in greater detail.