centralize accessing the agentIDE and related values (#5750)
Previously, we were doing 2 problematic things:

1. Determining extension/agent behavior based on the identity of the
editor (`agentIDE`), instead of based on the client's reported
capabilities. This is brittle and it is not clear *what* about (say) VS
Code makes this code path work and this one not.
2. Taking the editor's identity from the agent (but falling back to VS
Code implicitly), which meant that we were inconsistently doing
`agentIDE ?? CodyIDE.VSCode` and similar for the extension version in
many places. Of course, it's easy to make a mistake here.

This doesn't solve all of the problems, but it helps:

1. All places where we read the `agentIDE` are now centralized in a
global `clientCapabilities()`. This will make future refactoring easier,
solves a few problems (such as *explicitly* defaulting to
`CodyIDE.VSCode`, which is not a great idea but at least codifies our
current behavior with the type system), and makes it clear that this
value can't change throughout the lifetime of the extension.
2. The gotchas when reading `agentIDE` and related values are
documented, and the symbols are jsdoc'd with `@deprecated` to make them
noticeable.

There is no intentional user-facing behavior change.

I found this code tricky when improving the sign-in page in
https://github.com/sourcegraph/cody/pull/5746 because it was not clear
why something was gated for VS Code. For example, the JetBrains UI in
the storybook was configured to claim that it was running on the web,
which made no sense but might be because it needed a UI more similar to
what Cody Web used, and someone just reused that flag. We want to get
away from that!

## Test plan

Agent tests and e2e tests are pretty good here. Also inspect telemetry
and ensure that the right client name is being sent.