pass chat history to webview using webviewAPI, not old postMessage protocol (#5741)
Previously, chat history was synced to the webview using the old
postMessage protocol. Now, the ChatHistoryManager exposes an observable
and that is read directly (using the standard webviewAPI) by the
webview. This is a code refactor with no user-facing behavior change.

## Test plan

In Cody Web and in VS Code, try downloading chat history. Ensure the
JSON file is downloaded and is a JSON array of chat history.