Command: (fix) register missing commands when unifiedPrompts is enabled
Currently, some commands unrelated to the chat commands are not registered when unified prompts feature is enabled, which I believe is unintentional.

This PR add some documentations around some cody commands, and refactor the logic to make it clear what the new unified prompts commands are during the command registeration.

This PR also fixes the issue where the generate unit test command was unintentionally removed when the unified prompts feature is enabled.