use the accurate docContext and persist repo visibility using local storage (#5767)
The PR improves the following:
1. When calculating the `diff` over time for suggestion, we were using
`document.getText`. Analyzing offline telemetry we see in ~20% we get
some extra chars if user types between calculating `docContext` and
`loaded` call. We leverage already computed `completePrefix` and
`completeSuffix` to compute the text, which is accurate.
2. We were using in-memory cache for calculating `repoVisibility`
before. But without passing token, we frequently hit rate limit for
querying github api and cached tokens are evicted from in-memory cache
when user refresh. This PR stores the tokens in `localStorage` instead
of in-memory store to cache the repo visibility.
3. Adds the `code qwen` and `deepseek` model optimized for long prompts.

## Test plan
1. Added unit tests.
2. Manual Testing: Adding debugger and refreshing to see if the token
value is persisted/ deleted from local storage.