fix/chat: Wait for 'ready' before sending theme data. (#2371)
We used to inject themes using CSSOM injected by the client. We moved
this code to the Agent to share it with VS, etc. and used
DOMContentLoaded as the signal to do it. This races with webview React,
which may be responding to this signal by DOMContentLoaded, but is not
guaranteed to be.

Now we watch for the webview to send the 'ready' event before
dispatching the initial theme update message.

Fixes CODY-3796, BUGS-126, BUGS-233, BUGS-154.

## Test plan

Manually tested:

- Start and restart the IDE a bunch of times, the theme should be
applied.
- Switch the theme, that should still work.