Remove automatically generated Tokens to reduce redundant tokens (#5756)
Solves [Linear
issue](https://linear.app/sourcegraph/issue/CODY-350/bug-revoke-the-access-token-when-a-user-signs-out-of-jetbrains-and)

## Problem Statement
When signing in to Sourcegraph (and potentially JetBrains) using certain
sign-in providers, access tokens are automatically generated. These
tokens are subject to deletion under specific conditions, primarily upon
logging out. However, there are nuances depending on the method of
sign-in, the access token source, and whether multiple instances are
involved.

Details:
1. Automatic Token Generation:
•Tokens can be automatically generated when signing in using one of the
following providers:
•GitHub
•GitLab
•Google
•Sourcegraph instance (via  sign-in without token for v5.1 and above)
2. Token Deletion Criteria:
•An automatically generated token will only be deleted upon logging out
of the instance it was created for.
•If you sign out, the token will be deleted.
•However, if you switch instances or change access tokens while signed
in using an automatically generated token, the token will not be
deleted.
•Token deletion only occurs when the user logs out from the instance
where the token was generated, and only if the token was created using
one of the automatic sign-in methods listed above.
3. Instance and Token Behavior:
•If the user switches to a different instance or uses a different token
during a session, the original token remains active until explicitly
logged out.
•If the user returns to the original instance or token, no deletion
occurs unless the logout process is followed.
4. Manual Token Override:
•If the user manually inputs an access token from the instance (rather
than using automatic generation via sign-in), the token is not subject
to automatic deletion on logout.


[See video Proof
](https://www.loom.com/share/2a7bea8e2bbe4dcabc85cf7429a65c9d)

Thanks to Peter Guy, Dominic and Erik for the strategy on this.

## Test plan
Tested locally on both a local sourcegraph instance and with dotcom
instance using Vscode cody.


## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

---------

Co-authored-by: Dominic Cooney <dominic.cooney@sourcegraph.com>