Rename onebox telemetry metadata fields (#5770)
This PR renames a bunch of OneBox telemetry events metadata fields and
make them camelcase.

~~This PR also removes the check `isDotCom` from the `promptText`
private metadata field passed with the `chat-question:executed` event.
None of the private metadata fields are sent to SG through the
telemetery pipeline, so that data is still private, similar to the other
present privateMetadata field. The only change is that the `promptText`
will be recorded in the event_logs on that sourcegraph instance.~~

We need to remove the check so that we can selectively collect that
field for the S2 instance, configured by a env var here:
https://github.com/sourcegraph/cloud/pull/15143


## Test plan

Make sure all the updated events are still recorded correctly. 

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->