higher default timeout than 6s
The 6000ms timeout may be too short for some users on high-latency network connections. I have experienced this as a problem once or twice recently when on airplane (satellite) wifi. This may also help with https://github.com/sourcegraph/cody/issues/5783, but I am not certain of that since this timeout has been in place for a while and the reports of that issue are new.