Fix styling of avatar component when no avatarURL is present (#5811)
This is a stop-gap commit that fixes the users profile picture so that
it doesn't break the UI. The previous behavior can be viewed [in this
issue](https://linear.app/sourcegraph/issue/SRCH-1139/fix-chat-ui-issues-caused-by-profile-pictures-in-cody).

Ultimately, we should follow up with another PR with a more elegant
solution, but this fixes the immediate problem.

The issue here is that a height value of 'fit-content' will stretch to
fit everything within the container. This is fine when we're dealing
with a valid avatar link, but in the event that the link is broken, it
will default to the alt text, which is too long to maintain the uniform
size of the avatar component. Changing the value of height from
`fit-content` to `100%` is a quick fix.

Ultimately, we should model the UserAvatar component after Avatar
component from the sveltekit library. I'll do that in a follow-up PR
## Test plan
- Pass a broken avatar link into the UserAvatar component and make sure
it doesn't break the UI.
- Make sure that the new value for height does not break the component
if a valid url is passed in.
- Passing CI
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog
<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->
- Fixes a bug where avatar's were stretching to fit the content of the
alt text instead of remaining a static size.