fix agent tests model preferences state
In the agent tests, the `testing/reset` command was not entirely resetting the state between tests. This meant that 2 tests (explain and smell) that performed explicit `chat/setModel` requests to change the model to Mixtral were running with the chat model used for the previous tests.

The reason is that `testing/reset` just cleared the in-memory `LocalStorageProvider` backing `Map` and did not cause `LocalStorageProvider` to fire a change event, which is necessary for the `ModelsService` to know that the models preferences changed.

Also clean up types.