do not block on getting the site version or is-cody-enabled during auth (#5823)
The site version and Cody API version is now exposed in the
`siteVersion` global observable, which returns the site version and Cody
API version for the currently authenticated endpoint (and caches the
value).

If the Sourcegraph instance does not have Cody enabled, this means that
the error message will come later, not at sign-in. With one box, the
lines are getting blurrier anyway, and the logic for determining if Cody
was enabled was complex and often treated network errors as "Cody is
disabled". It is not worth it to invest time in making this more robust
because few users will encounter this, and those who do will just get an
error that Cody is disabled when they actually try to perform any
operation.

This means that initial auth takes 1-3 less HTTP requests.

## Test plan

CI & e2e